refactor: organize leaf agent interactions
