+++ apps/leaf/src/internal/agentRuntime/actions/submitAgentInput/agentInputNotes.ts
@@ -1,0 +1,2 @@
+export const QUEUED_TURN_WITHDRAWAL_NOTE =
+	"(The user sent a newer message before this was shown, so it was withdrawn. Do not rebuild or ask anything — reply with nothing; act on the user's next message instead.)";
