+++ apps/leaf/src/internal/agentRuntime/actions/submitAgentInput/types.ts
@@ -1,0 +1,12 @@
+import type {
+	ChainedPendingRequest,
+	PendingQuestion,
+} from "../../eve/parkedInput.js";
+
+export type ResumedAgentTurn = Readonly<{
+	chained?: ChainedPendingRequest;
+	chainedSiblingRequestIds: ReadonlyArray<string>;
+	deferredEmptyTurn: boolean;
+	question?: PendingQuestion;
+	text: string;
+}>;
